// ui/src/services/cradle/models/StatisticsNote.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * CRADLE
 * Threat Intelligence Knowledge Management
 *
 * The version of the OpenAPI document: 2.10.2-beta.9bd46310 (2.10.2-beta.9bd46310)
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { StatisticsUser } from './StatisticsUser';
import {
    StatisticsUserFromJSON,
    StatisticsUserFromJSONTyped,
    StatisticsUserToJSON,
    StatisticsUserToJSONTyped,
} from './StatisticsUser';

/**
 * Serializer for notes in statistics
 * @export
 * @interface StatisticsNote
 */
export interface StatisticsNote {
    /**
     * 
     * @type {string}
     * @memberof StatisticsNote
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof StatisticsNote
     */
    readonly content?: string;
    /**
     * 
     * @type {string}
     * @memberof StatisticsNote
     */
    title: string;
    /**
     * 
     * @type {Date}
     * @memberof StatisticsNote
     */
    timestamp: Date;
    /**
     * 
     * @type {StatisticsUser}
     * @memberof StatisticsNote
     */
    author: StatisticsUser;
}

/**
 * Check if a given object implements the StatisticsNote interface.
 */
export function instanceOfStatisticsNote(value: object): value is StatisticsNote {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('timestamp' in value) || value['timestamp'] === undefined) return false;
    if (!('author' in value) || value['author'] === undefined) return false;
    return true;
}

export function StatisticsNoteFromJSON(json: any): StatisticsNote {
    return StatisticsNoteFromJSONTyped(json, false);
}

export function StatisticsNoteFromJSONTyped(json: any, ignoreDiscriminator: boolean): StatisticsNote {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'content': json['content'] == null ? undefined : json['content'],
        'title': json['title'],
        'timestamp': (new Date(json['timestamp'])),
        'author': StatisticsUserFromJSON(json['author']),
    };
}

export function StatisticsNoteToJSON(json: any): StatisticsNote {
    return StatisticsNoteToJSONTyped(json, false);
}

export function StatisticsNoteToJSONTyped(value?: Omit<StatisticsNote, 'content'> | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'timestamp': ((value['timestamp']).toISOString()),
        'author': StatisticsUserToJSON(value['author']),
    };
}